+++ R/future-options.R
@@ -1,157 +1,333 @@
-#' `future` specific options
+#' Options to fine tune furrr
 #'
-#' These options are used by `future()` internally to tweak the environment that
-#' the expressions are called in. The most important ones are `globals` and
-#' `packages` which allow you to be explicit about the variables and packages
-#' that are exported to each worker.
+#' These options fine tune furrr functions, such as [future_map()]. They
+#' are either used by furrr directly, or are passed on to [future::future()].
 #'
-#' @param globals A logical, a character vector, or a named list for
-#'        controlling how globals are handled. For details, see `Global variables and packages`.
+#' @param ... These dots are reserved for future extensibility and must
+#'   be empty.
 #'
-#' @param packages (optional) a character vector specifying packages
-#'        to be attached in the R environment evaluating the future.
+#' @param stdout A logical.
 #'
-#' @param seed A logical or an integer (of length one or seven),
-#'        or a list of `length(.x)` with pre-generated random seeds.
-#'        For details, see below section.
+#'   - If `TRUE`, standard output of the underlying futures is captured and
+#'     relayed as soon as possible.
 #'
-#' @param lazy Specifies whether the futures should be resolved
-#'        lazily or eagerly (default).
+#'   - If `FALSE`, output is silenced by sinking it to the null device.
 #'
-#' @param scheduling Average number of futures ("chunks") per worker.
-#'        If `0.0`, then a single future is used to process all elements
-#'        of `.x`.
-#'        If `1.0` or `TRUE`, then one future per worker is used.
-#'        If `2.0`, then each worker will process two futures
-#'        (if there are enough elements in `.x`).
-#'        If `Inf` or `FALSE`, then one future per element of
-#'        `.x` is used.
+#'   - If `NA`, output is not intercepted. This is not recommended.
 #'
-#' @section Global variables and packages:
+#' @param conditions A character string of conditions classes to be captured
+#'   and relayed. The default is the same as the condition argument of
+#'   [future::Future()]. To not intercept conditions, use
+#'   `conditions = character(0L)`. Errors are always relayed.
 #'
-#' By default, the `future` package will perform black magic to look up the
-#' global variables and packages that your `furrr` call requires, and it
-#' will export these to each worker. However, it is not always perfect, and
-#' can be refined with the `globals` and `packages` arguments.
+#' @param globals A logical, a character vector, a named list, or `NULL` for
+#'   controlling how globals are handled. For details, see `Global variables`.
 #'
-#' `globals` may be used to control how globals
-#' should be handled similarly how the `globals` argument is used with
-#' `future()`. Since all function calls use the same set of globals, this function can do
-#' any gathering of globals upfront (once), which is more efficient than if
-#' it would be done for each future independently.
+#' @param packages A character vector, or `NULL`. If supplied, packages that
+#'   are guaranteed to be attached in the R environment where the future
+#'   is evaluated.
 #'
-#' * If `TRUE` or `NULL`, then globals are automatically identified and gathered.
-#' * If a character vector of names is specified, then those globals are gathered.
+#' @param lazy A logical. Specifies whether futures should be resolved
+#'   lazily or eagerly.
+#'
+#' @param seed A logical, an integer of length `1` or `7`, or a list of
+#'   `length(.x)` with pre-generated random seeds. For details, see the
+#'   `Reproducible random number generation (RNG)` section.
+#'
+#' @param scheduling A single integer, logical, or `Inf`. This argument
+#'   controls the average number of futures ("chunks") per worker.
+#'
+#'   - If `0`, then a single future is used to process all elements of `.x`.
+#'
+#'   - If `1` or `TRUE`, then one future per worker is used.
+#'
+#'   - If `2`, then each worker will process two futures (provided there
+#'     are enough elements in `.x`).
+#'
+#'   - If `Inf` or `FALSE`, then one future per element of `.x` is used.
+#'
+#'   Only used if `chunk_size` is `NULL`.
+#'
+#' @param chunk_size A single integer, `Inf`, or `NULL`. This argument
+#'   controls the average number of elements per future (`"chunk"`). If `Inf`,
+#'   then all elements are processed in a single future. If `NULL`, then
+#'   `scheduling` is used instead to determine how `.x` is chunked.
+#'
+#' @param prefix A single character string, or `NULL`. If a character string,
+#'   then each future is assigned a label as `{prefix}-{chunk-id}`. If `NULL`,
+#'   no labels are used.
+#'
+#'
+#' @section Global variables:
+#'
+#' `globals` controls how globals are identified, similar to the `globals`
+#' argument of [future::future()]. Since all function calls use the same set of
+#' globals, furrr gathers globals upfront (once), which is more efficient than
+#' if it was done for each future independently.
+#'
+#' * If `TRUE` or `NULL`, then globals are automatically identified and
+#'   gathered.
+#'
+#' * If a character vector of names is specified, then those globals are
+#'   gathered.
+#'
 #' * If a named list, then those globals are used as is.
-#' * In all cases, `.f` and any `...` arguments are automatically
-#' passed as globals to each future created as they are always needed.
 #'
-#' `packages` may be used to control the packages that are exported
-#' to each worker.
-#'
-#' * If a character vector of packages names is specified, those are exported
-#' to each worker.
-#' * In all cases, `purrr` is exported, as it is always required on each worker.
+#' * In all cases, `.f` and any `...` arguments are automatically passed as
+#'   globals to each future created, as they are always needed.
 #'
 #' @section Reproducible random number generation (RNG):
 #'
-#' Unless `seed = FALSE`, this function guarantees to generate
+#' Unless `seed = FALSE`, furrr functions are guaranteed to generate
 #' the exact same sequence of random numbers _given the same initial
-#' seed / RNG state_ - this regardless of type of futures and scheduling
+#' seed / RNG state_ regardless of the type of futures and scheduling
 #' ("chunking") strategy.
 #'
-#' RNG reproducibility is achieved by pregenerating the random seeds for all
-#' iterations (over `.x`) by using L'Ecuyer-CMRG RNG streams.  In each
-#' iteration, these seeds are set before calling \code{.f(.x[[ii]], ...)}.
+#' RNG reproducibility is achieved by pre-generating the random seeds for all
+#' iterations (over `.x`) by using L'Ecuyer-CMRG RNG streams. In each
+#' iteration, these seeds are set before calling `.f(.x[[i]], ...)`.
 #' _Note, for large `length(.x)` this may introduce a large overhead._
 #'
-#' As input (`seed`), a fixed seed (integer) may be given, either
-#' as a full L'Ecuyer-CMRG RNG seed (vector of 1+6 integers) or as a seed
-#' generating such a full L'Ecuyer-CMRG seed.
-#' If `seed = TRUE`, then \code{\link[base:Random]{.Random.seed}}
-#' is returned if it holds a L'Ecuyer-CMRG RNG seed, otherwise one is created
-#' randomly.
+#' A fixed `seed` may be given as an integer vector, either as a full
+#' L'Ecuyer-CMRG RNG seed of length `7`, or as a seed of length `1` that
+#' will be used to generate a full L'Ecuyer-CMRG seed.
+#'
+#' If `seed = TRUE`, then `.Random.seed` is returned if it holds a
+#' L'Ecuyer-CMRG RNG seed, otherwise one is created randomly.
+#'
 #' If `seed = NA`, a L'Ecuyer-CMRG RNG seed is randomly created.
-#' If none of the function calls \code{.f(.x[[ii]], ...)} uses random number
+#'
+#' If none of the function calls `.f(.x[[i]], ...)` use random number
 #' generation, then `seed = FALSE` may be used.
 #'
 #' In addition to the above, it is possible to specify a pre-generated
-#' sequence of RNG seeds as a list such that
-#' `length(seed) == length(.x)` and where each element is an
-#' integer seed that can be assigned to \code{\link[base:Random]{.Random.seed}}.
-#' Use this alternative with caution.
-#' **Note that `as.list(seq_along(.x))` is _not_ a valid set of such
-#' `.Random.seed` values.**
+#' sequence of RNG seeds as a list such that `length(seed) == length(.x)` and
+#' where each element is an integer seed that can be assigned to `.Random.seed`.
+#' Use this alternative with caution. **Note that `as.list(seq_along(.x))` is
+#' _not_ a valid set of such `.Random.seed` values.**
 #'
-#' In all cases but `seed = FALSE`, the RNG state of the calling
-#' R processes after this function returns is guaranteed to be
-#' "forwarded one step" from the RNG state that was before the call and
-#' in the same way regardless of `seed`, `scheduling`
-#' and future strategy used.  This is done in order to guarantee that an \R
+#' In all cases but `seed = FALSE`, after a furrr function returns, the RNG
+#' state of the calling R process is guaranteed to be "forwarded one step" from
+#' the RNG state before the call. This is true regardless of the future
+#' strategy / scheduling used. This is done in order to guarantee that an R
 #' script calling `future_map()` multiple times should be numerically
 #' reproducible given the same initial seed.
 #'
 #' @export
-future_options <- function(globals = TRUE, packages = NULL, seed = FALSE, lazy = FALSE, scheduling = 1.0) {
+#' @examples
+#' furrr_options()
+furrr_options <- function(...,
+                          stdout = TRUE,
+                          conditions = NULL,
+                          globals = TRUE,
+                          packages = NULL,
+                          lazy = FALSE,
+                          seed = FALSE,
+                          scheduling = 1.0,
+                          chunk_size = NULL,
+                          prefix = NULL) {
+  ellipsis::check_dots_empty()
 
-  .options <- new_future_options(
-    globals    = globals,
-    packages   = packages,
-    seed       = seed,
-    lazy       = lazy,
+  stdout <- validate_stdout(stdout)
+  conditions <- validate_conditions(conditions)
+  globals <- validate_globals(globals)
+  packages <- validate_packages(packages)
+  lazy <- validate_lazy(lazy)
+  seed <- validate_seed(seed)
+  scheduling <- validate_scheduling(scheduling)
+  chunk_size <- validate_chunk_size(chunk_size)
+  prefix <- validate_prefix(prefix)
+
+  out <- list(
+    stdout = stdout,
+    conditions = conditions,
+    globals = globals,
+    packages = packages,
+    lazy = lazy,
+    seed = seed,
+    scheduling = scheduling,
+    chunk_size = chunk_size,
+    prefix = prefix
+  )
+
+  structure(out, class = "furrr_options")
+}
+
+#' @export
+print.furrr_options <- function(x, ...) {
+  cat("<furrr_options>")
+}
+
+# ------------------------------------------------------------------------------
+
+#' Deprecated furrr options
+#'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `future_options()` has been deprecated in favor of [furrr_options()] as
+#' of furrr 0.2.0.
+#'
+#' @inheritParams furrr_options
+#'
+#' @keywords internal
+#' @export
+#' @examples
+#' future_options()
+future_options <- function(globals = TRUE,
+                           packages = NULL,
+                           seed = FALSE,
+                           lazy = FALSE,
+                           scheduling = 1.0) {
+  lifecycle::deprecate_warn("0.2.0", "future_options()", "furrr_options()")
+
+  furrr_options(
+    globals = globals,
+    packages = packages,
+    seed = seed,
+    lazy = lazy,
     scheduling = scheduling
   )
-
-  validate_future_options(.options)
 }
 
-new_future_options <- function(globals, packages, seed, lazy, scheduling) {
+# ------------------------------------------------------------------------------
 
-  stopifnot(inherits(globals,    c("NULL", "logical", "character", "list")))
-  stopifnot(inherits(packages,   c("NULL", "character")))
-  stopifnot(inherits(seed,       c("logical", "integer", "list")))
-  stopifnot(inherits(lazy,       c("logical")))
-  stopifnot(inherits(scheduling, c("logical", "numeric")))
+assert_furrr_options <- function(x) {
+  if (!is_furrr_options(x)) {
+    abort("`.options` must be created from `furrr_options()`.")
+  }
 
-  structure(
-    list(
-      globals    = globals,
-      packages   = packages,
-      seed       = seed,
-      lazy       = lazy,
-      scheduling = scheduling
-    ),
-    class = "future_options"
-  )
+  invisible(x)
 }
 
-validate_future_options <- function(.options) {
+is_furrr_options <- function(x) {
+  inherits(x, "furrr_options")
+}
 
-  # Globals checks
-  if(is.list(.options$globals)) {
-    nms <- names(.options$globals)
-    if(any(nms == "") || is.null(nms)) {
-      stop("Every element of a globals list must be named.", call. = FALSE)
+# ------------------------------------------------------------------------------
+
+validate_stdout <- function(x) {
+  vctrs::vec_assert(x, size = 1L, arg = "stdout")
+  vctrs::vec_cast(x, logical(), x_arg = "stdout")
+}
+
+validate_conditions <- function(x) {
+  vctrs::vec_cast(x, character(), x_arg = "conditions")
+}
+
+validate_globals <- function(x) {
+  if (is.null(x)) {
+    return(TRUE)
+  }
+
+  if (is.logical(x)) {
+    return(x)
+  }
+
+  if (is.character(x)) {
+    return(x)
+  }
+
+  if (is.list(x)) {
+    if (!is_named(x)) {
+      abort("`globals` cannot be an unnamed or partially named list.")
     }
+
+    return(x)
   }
 
-  # Seed checks
-  if(is.integer(.options$seed)) {
-    if( !(length(.options$seed) == 1L || length(.options$seed) == 7L) ) {
-      stop("Integer seeds must be of length 1 or 7.", call. = FALSE)
-    }
-  }
-
-  # Scheduling checks
-  if(!length(.options$scheduling == 1L)) {
-    stop("scheduling must be length 1.", call. = FALSE)
-  }
-
-  .options
+  abort("`globals` must be `NULL`, a logical, a character vector, or a named list.")
 }
 
-assert_future_options <- function(.options) {
-  if(!inherits(.options, "future_options")) {
-    stop(".options must be created from future_options().", call. = FALSE)
+validate_packages <- function(x) {
+  vctrs::vec_cast(x, character(), x_arg = "packages")
+}
+
+validate_lazy <- function(x) {
+  vctrs::vec_assert(x, size = 1, arg = "lazy")
+  x <- vctrs::vec_cast(x, logical(), x_arg = "lazy")
+
+  if (is.na(x)) {
+    abort("`lazy` can't be `NA`.")
   }
+
+  x
+}
+
+validate_seed <- function(x) {
+  # Size is validated when we have `.x`
+  if (is.list(x)) {
+    return(x)
+  }
+
+  if (is.logical(x)) {
+    if (length(x) != 1L) {
+      abort("A logical `seed` must be length 1.")
+    }
+
+    return(x)
+  }
+
+  if (is.double(x)) {
+    x <- vctrs::vec_cast(x, integer(), x_arg = "seed")
+  }
+
+  if (is.integer(x)) {
+    if (length(x) != 1L && length(x) != 7L) {
+      abort("An integer `seed` must be length 1 or 7.")
+    }
+
+    if (any(is.na(x))) {
+      abort("An integer `seed` cannot have `NA` values.")
+    }
+
+    return(x)
+  }
+
+  abort("`seed` must be a logical, integer, or list.")
+}
+
+validate_scheduling <- function(x) {
+  if (length(x) != 1L) {
+    abort("`scheduling` must be length 1.")
+  }
+
+  if (identical(x, Inf)) {
+    return(x)
+  }
+
+  if (is.logical(x)) {
+    if (!is_bool(x)) {
+      abort("A logical `scheduling` value can't be `NA`.")
+    }
+
+    return(x)
+  }
+
+  vctrs::vec_cast(x, integer(), x_arg = "scheduling")
+}
+
+validate_chunk_size <- function(x) {
+  if (is.null(x)) {
+    return(x)
+  }
+
+  if (identical(x, Inf)) {
+    return(x)
+  }
+
+  vctrs::vec_assert(x, size = 1L, arg = "chunk_size")
+  vctrs::vec_cast(x, integer(), arg = "chunk_size")
+}
+
+validate_prefix <- function(x) {
+  if (is.null(x)) {
+    return(x)
+  }
+
+  if (!is_string(x)) {
+    abort("`prefix` must be a character string, or `NULL`.")
+  }
+
+  x
 }