+++ R/future-pmap.R
@@ -1,6 +1,6 @@
 #' @rdname future_map2
 #' @export
-future_pmap <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap")
 
   if (is.data.frame(.l)) {
@@ -12,7 +12,7 @@ future_pmap <- function(.l, .f, ..., .options = future_options(), .progress = de
 
 #' @rdname future_map2
 #' @export
-future_pmap_chr <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap_chr <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_chr")
 
   if (is.data.frame(.l)) {
@@ -24,7 +24,7 @@ future_pmap_chr <- function(.l, .f, ..., .options = future_options(), .progress 
 
 #' @rdname future_map2
 #' @export
-future_pmap_dbl <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap_dbl <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_dbl")
 
   if (is.data.frame(.l)) {
@@ -36,7 +36,7 @@ future_pmap_dbl <- function(.l, .f, ..., .options = future_options(), .progress 
 
 #' @rdname future_map2
 #' @export
-future_pmap_int <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap_int <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_int")
 
   if (is.data.frame(.l)) {
@@ -48,7 +48,7 @@ future_pmap_int <- function(.l, .f, ..., .options = future_options(), .progress 
 
 #' @rdname future_map2
 #' @export
-future_pmap_lgl <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap_lgl <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_lgl")
 
   if (is.data.frame(.l)) {
@@ -60,7 +60,7 @@ future_pmap_lgl <- function(.l, .f, ..., .options = future_options(), .progress 
 
 #' @rdname future_map2
 #' @export
-future_pmap_dfr <- function(.l, .f, ..., .id = NULL, .options = future_options(), .progress = deprecated()) {
+future_pmap_dfr <- function(.l, .f, ..., .id = NULL, .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_dfr")
 
   # Passing through the template doesn't work because of the way fold() works.
@@ -75,7 +75,7 @@ future_pmap_dfr <- function(.l, .f, ..., .id = NULL, .options = future_options()
 
 #' @rdname future_map2
 #' @export
-future_pmap_dfc <- function(.l, .f, ..., .options = future_options(), .progress = deprecated()) {
+future_pmap_dfc <- function(.l, .f, ..., .options = furrr_options(), .progress = deprecated()) {
   maybe_warn_deprecated_progress(is_present(.progress), what = "future_pmap_dfc")
 
   # Passing through the template doesn't work because of the way fold() works.