+++ tests/testthat/test-deprecation.R
@@ -43,3 +43,20 @@ furrr_test_that("`.progress` updates are deprecated", {
   expect_deprecated(future_modify_at("1", 1, ~.x, .progress = TRUE), "future_modify_at")
   expect_deprecated(future_modify_if(1, TRUE, ~.x, .progress = TRUE), "future_modify_if")
 })
+
+test_that("can use deprecated `future_options()`", {
+  expect_identical(
+    expect_deprecated(
+      future_options(),
+      "`future_options[(][)]` is deprecated as of furrr 0.2.0"
+    ),
+    furrr_options()
+  )
+
+  expect_identical(
+    expect_deprecated(
+      future_options(globals = "x", packages = "dplyr", seed = 1, lazy = TRUE, scheduling = 2)
+    ),
+    furrr_options(globals = "x", packages = "dplyr", seed = 1, lazy = TRUE, scheduling = 2)
+  )
+})